+++ timeutil/timeutil_test.go
@@ -391,6 +391,7 @@ func (s *TimeUtilSuite) TestHelpers(c *C) {
 	c.Assert(AddWorkdays(d, -10).String(), Equals, "2012-05-18 12:00:00 +0000 UTC")
 
 	c.Assert(ToSeconds(60), Equals, time.Minute)
+	c.Assert(ToSeconds(60.5).Seconds(), Equals, 60.5)
 	c.Assert(ToMinutes(60), Equals, time.Hour)
 	c.Assert(ToHours(1), Equals, time.Hour)
 	c.Assert(ToDays(1), Equals, 24*time.Hour)